notes on GameEngine
created ToDo. We need to make it possible to continue the game if player selects a wrong playcount; that is, first make them select a legal playcount number. so the illegalargument exception shouldn't print to the console - rather, we should have a while loop with a try-catch block, or some other means of looping back to selcting play count when the argument is invalid