add while loop and try/catch block to HumanPlayer
human player should only type in the name of a valid MoveSelection (we'll parse to caps, we're not monsters, they don't have to type in all caps). The console should prompt them again in the event they input an invalid argument, rather than crashing and exiting the game. This is accomplished with the if/else statement nested in a try/catch, nested in a while loop. Currently I still have it crashing, because I'll need to create a custom exception for this instance